Removed EntraGroup, Added EntraGroupOwner and EntraGroupAppRoleAssignment, Updated EntraMSLifecyclePolicyGroup test

+++ test/module/Entra/Integration/EntraGroupAppRoleAssignment.Tests.ps1
@@ -10,26 +10,31 @@ Describe "The EntraGroup command executing unmocked" {
             Connect-Entra -TenantId $tenantId  -AppId $appId -CertificateThumbprint $cert
 
             $thisTestInstanceId = New-Guid | Select-Object -expandproperty guid
-            $testName = 'Demo Name' + $thisTestInstanceId
-
+            $global:displayName = 'DemoName' + $thisTestInstanceId
+            
             #create test group 
-            $global:newGroup = New-EntraGroup -DisplayName $testName -MailEnabled $false -SecurityEnabled $true -MailNickName "NotSet" 
+            $global:newGroup = New-EntraGroup -DisplayName $displayName -MailEnabled $false -SecurityEnabled $true -MailNickName  $testName
+            Write-Host $newGroup
         }
 
         It "should successfully get a specific group by using an Id" {
             $group = Get-EntraGroup -ObjectId $newGroup.Id
             $group.Id | Should -Be $newGroup.Id
-            $group.DisplayName | Should -Be $testName
+            $group.DisplayName | Should -Be $displayName
         }
 
         It "should successfully update a group display name" {
-            Set-EntraGroup -ObjectId $newGroup.Id -DisplayName "Demo Name 2"
+            $global:updatedDisplayName = "Demo Name 2"
+            Set-EntraGroup -ObjectId $newGroup.Id -DisplayName $updatedDisplayName
             $result = Get-EntraGroup -ObjectId $newGroup.Id
             $result.Id | Should -Contain $newGroup.Id
         }
 
+        
         AfterAll {
-            Remove-EntraGroup -ObjectId $newGroup.Id | Out-Null
+            if ($newGroup) {
+                Remove-EntraGroup -ObjectId $newGroup.Id | Out-Null
+            }
         }
     }
 }
